fix I32LP64 compile
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@39388 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ tools/switchboard2/request_thread.cc
@@ -274,7 +274,7 @@ request_thread::handle_normal_request(fcgi::record &initial)
 	dend = newparams.end();
 
 	while (dit != dend) {
-		int now = std::min(65535, std::distance(dit, dend));
+		int now = std::min(65535, (int) std::distance(dit, dend));
 		r_params.contentData.resize(now);
 		std::copy(dit, dit + now, r_params.contentData.begin());
 		r_params.content_length(now);
@@ -300,7 +300,7 @@ request_thread::handle_normal_request(fcgi::record &initial)
 	 * Only 64K in allowed in one record.
 	 */
 	while (dit != dend) {
-		int now = std::min(65535, std::distance(dit, dend));
+		int now = std::min(65535, (int) std::distance(dit, dend));
 		r_stdin.contentData.resize(now);
 		std::copy(dit, dit + now, r_stdin.contentData.begin());
 		r_stdin.content_length(now);
